Fix for a number of issues found in org.campagnelab.ui 1.0:
FileSelectionButton generator must use the property that indicates if the selection should be files or dirs. It did not in 1.0.
node statements in call back function were not initialized for FileSelection buttons.
The type of node in the call back concept function was not determined, forcing the end user to cast to the correct concept.